// frontend/src/app/generated/model/note.ts
/**
 * Thinkify - OpenAPI 3.0
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Note { 
    id?: number;
    createdAt?: string;
    updatedAt?: string;
    name?: string;
    rawText?: string;
    aiOutput?: string;
}